requirements.txt
pandas
scrapy
googlesearch-python